Move websocket spring classes into websocket package
